Make small refactors
