+++ app/server/src/main/scala/org/bitcoins/server/BitcoinSServerMain.scala
@@ -14,12 +14,7 @@ import org.bitcoins.commons.serializers.WsPicklers
 import org.bitcoins.commons.util.{DatadirParser, ServerArgParser}
 import org.bitcoins.core.api.chain.ChainApi
 import org.bitcoins.core.api.feeprovider.FeeRateApi
-import org.bitcoins.core.api.node.{
-  ExternalImplementationNodeType,
-  InternalImplementationNodeType,
-  NodeApi,
-  NodeType
-}
+import org.bitcoins.core.api.node.{ExternalImplementationNodeType, InternalImplementationNodeType, NodeApi, NodeType}
 import org.bitcoins.core.protocol.transaction.Transaction
 import org.bitcoins.core.util.{NetworkUtil, TimeUtil}
 import org.bitcoins.core.wallet.fee.SatoshisPerVirtualByte
@@ -36,12 +31,7 @@ import org.bitcoins.rpc.BitcoindException.InWarmUp
 import org.bitcoins.rpc.client.common.BitcoindRpcClient
 import org.bitcoins.rpc.config.{BitcoindRpcAppConfig, ZmqConfig}
 import org.bitcoins.server.routes.{BitcoinSServerRunner, CommonRoutes, Server}
-import org.bitcoins.server.util.{
-  BitcoinSAppScalaDaemon,
-  ChainUtil,
-  ServerBindings,
-  WsServerConfig
-}
+import org.bitcoins.server.util.{BitcoinSAppScalaDaemon, ChainUtil, ServerBindings, WebsocketUtil, WsServerConfig}
 import org.bitcoins.tor.config.TorAppConfig
 import org.bitcoins.wallet._
 import org.bitcoins.wallet.config.WalletAppConfig
@@ -184,7 +174,7 @@ class BitcoinSServerMain(override val serverArgParser: ServerArgParser)(implicit
                                 wallet = wallet,
                                 dlcNode = dlcNode,
                                 serverCmdLineArgs = serverArgParser)
-      walletCallbacks = buildWalletCallbacks(server.walletQueue, chainApi)
+      walletCallbacks = WebsocketUtil.buildWalletCallbacks(server.walletQueue, chainApi)
       _ = walletConf.addCallbacks(walletCallbacks)
       _ = {
         logger.info(
@@ -267,7 +257,7 @@ class BitcoinSServerMain(override val serverArgParser: ServerArgParser)(implicit
                                 wallet = wallet,
                                 dlcNode = dlcNode,
                                 serverCmdLineArgs = serverArgParser)
-      walletCallbacks = buildWalletCallbacks(server.walletQueue, bitcoind)
+      walletCallbacks = WebsocketUtil.buildWalletCallbacks(server.walletQueue, bitcoind)
       _ = walletConf.addCallbacks(walletCallbacks)
     } yield {
       logger.info(s"Done starting Main!")
@@ -520,88 +510,6 @@ class BitcoinSServerMain(override val serverArgParser: ServerArgParser)(implicit
     f
   }
 
-  private def buildWalletCallbacks(
-      walletQueue: SourceQueueWithComplete[Message],
-      chainApi: ChainApi): WalletCallbacks = {
-    val onAddressCreated: OnNewAddressGenerated = { addr =>
-      val notification = WalletNotification.NewAddressNotification(addr)
-      val json =
-        upickle.default.writeJs(notification)(WsPicklers.newAddressPickler)
-      val msg = TextMessage.Strict(json.toString())
-      val offerF = walletQueue.offer(msg)
-      offerF.map(_ => ())
-    }
-
-    val onTxProcessed: OnTransactionProcessed = { tx =>
-      buildTxNotification(wsType = WalletWsType.TxProcessed,
-                          tx = tx,
-                          walletQueue = walletQueue)
-    }
-
-    val onTxBroadcast: OnTransactionBroadcast = { tx =>
-      buildTxNotification(wsType = WalletWsType.TxBroadcast,
-                          tx = tx,
-                          walletQueue = walletQueue)
-    }
-
-    val onReservedUtxo: OnReservedUtxos = { utxos =>
-      val notification =
-        WalletNotification.ReservedUtxosNotification(utxos)
-      val notificationJson =
-        upickle.default.writeJs(notification)(WsPicklers.reservedUtxosPickler)
-      val msg = TextMessage.Strict(notificationJson.toString())
-      val offerF = walletQueue.offer(msg)
-      offerF.map(_ => ())
-    }
-
-    val onBlockProcessed: OnBlockProcessed = { block =>
-      val resultF =
-        ChainUtil.getBlockHeaderResult(block.blockHeader.hashBE, chainApi)
-      val f = for {
-        result <- resultF
-        notification =
-          WalletNotification.BlockProcessedNotification(result)
-        notificationJson =
-          upickle.default.writeJs(notification)(
-            WsPicklers.blockProcessedPickler)
-        msg = TextMessage.Strict(notificationJson.toString())
-        _ <- walletQueue.offer(msg)
-      } yield {
-        ()
-      }
-
-      f
-    }
-
-    WalletCallbacks(
-      onTransactionProcessed = Vector(onTxProcessed),
-      onNewAddressGenerated = Vector(onAddressCreated),
-      onReservedUtxos = Vector(onReservedUtxo),
-      onTransactionBroadcast = Vector(onTxBroadcast),
-      onBlockProcessed = Vector(onBlockProcessed)
-    )
-  }
-
-  private def buildTxNotification(
-      wsType: WalletWsType,
-      tx: Transaction,
-      walletQueue: SourceQueueWithComplete[Message]): Future[Unit] = {
-    val json = wsType match {
-      case WalletWsType.TxProcessed =>
-        val notification = WalletNotification.TxProcessedNotification(tx)
-        upickle.default.writeJs(notification)(WsPicklers.txProcessedPickler)
-      case WalletWsType.TxBroadcast =>
-        val notification = WalletNotification.TxBroadcastNotification(tx)
-        upickle.default.writeJs(notification)(WsPicklers.txBroadcastPickler)
-      case x @ (WalletWsType.NewAddress | WalletWsType.ReservedUtxos |
-          WalletWsType.BlockProcessed) =>
-        sys.error(s"Cannot build tx notification for $x")
-    }
-
-    val msg = TextMessage.Strict(json.toString())
-    val offerF = walletQueue.offer(msg)
-    offerF.map(_ => ())
-  }
 }
 
 object BitcoinSServerMain extends BitcoinSAppScalaDaemon {
